requirements.txt
streamlit>=1.37.0
Pillow
numpy
easyocr
opencv-python-headless
gspread
gspread-dataframe
google-auth
google-auth-oauthlib
google-api-python-client
python-dateutil
